+++ wheelodex/util.py
@@ -5,6 +5,7 @@ import re
 from   flask             import Response
 from   flask.json        import dumps
 from   packaging.version import parse
+import pyrfc3339
 import requests
 import requests_download
 from   wheel_inspect     import parse_wheel_filename
@@ -234,3 +235,17 @@ def glob2like(s):
         else:
             return '\\' + x
     return re.sub(r'(\x5C.|[?*%_])', subber, s)
+
+def parse_timestamp(s):
+    """ Parse an ISO 8601 timestamp, assuming anything naïve is in UTC """
+    if re.fullmatch(r'\d{4}-\d\d-\d\d[T ]\d\d:\d\d:\d\d(\.\d+)?', s):
+        s += 'Z'
+    return pyrfc3339.parse(s)
+
+    # Python 3.7+:
+    #if s.endswith('Z'):
+    #    s = s[:-1] + '+00:00'
+    #dt = datetime.fromisoformat(s)
+    #if dt.tzinfo is None:
+    #    dt = dt.replace(tzinfo=timezone.utc)
+    #return dt